+++ cmfgpu/phys/cuda/levee.cu
@@ -431,7 +431,7 @@ void launch_levee_bif_outflow(
     at::Tensor water_surface_elevation_ptr,
     at::Tensor protected_water_surface_elevation_ptr,
     at::Tensor total_storage_ptr, at::Tensor outgoing_storage_ptr,
-    float gravity, at::Tensor time_step_ptr,
+    double gravity, at::Tensor time_step_ptr,
     long num_catchments, long num_bifurcation_paths,
     int num_bifurcation_levels,
     long BLOCK_SIZE)
